adding option to cache release build

+++ install.sh
@@ -1,10 +1,31 @@
-#!/bin/sh
+#!/bin/bash
+
 BUILD_DIRECTORY="build-release"
-rm -rf $BUILD_DIRECTORY
-mkdir $BUILD_DIRECTORY
+CACHE_BUILD=0
+# Parse arguments
+for arg in "$@"; do
+  case $arg in
+    --cache)
+      CACHE_BUILD=1
+      shift # Remove argument from the list
+      ;;
+    *)
+      ;;
+  esac
+done
+
+# Optionally remove the build directory
+if [ $CACHE_BUILD -eq 0 ]; then
+  rm -rf $BUILD_DIRECTORY
+fi
+
+mkdir -p $BUILD_DIRECTORY
 cd $BUILD_DIRECTORY
 cmake -DCMAKE_BUILD_TYPE=Release -DFISHNET_TEST=OFF -DFISHNET_APPS=ON -DCMAKE_INSTALL_LIBDIR=lib -DCMAKE_INSTALL_PREFIX=$HOME/.local ..
 cmake --build . --config Release -j 16
 cmake --install .
 cd ..
-rm -rf $BUILD_DIRECTORY
+# Optionally clean up the build directory
+if [ $CACHE_BUILD -eq 0 ]; then
+  rm -rf $BUILD_DIRECTORY
+fi
